requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
ephem==4.2
fonttools==4.58.0
future==1.0.0
geocoder==1.38.1
gpsd-py3==0.3.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.2.5
packaging==25.0
pillow==11.2.1
pyephem==9.99
pyparsing==3.2.3
python-dateutil==2.9.0.post0
ratelim==0.1.6
requests==2.32.3
six==1.17.0
urllib3==2.4.0
